Python/bittorrent_traffic_detector/requirements.txt
bencode.py==4.0.0
scapy==2.4.5
simple_dht_node==1.2.2
